Move Admin::CurrencyHelper to CurrencyHelper
To be able to use it in Storefront as well
