+++ backend/app/controllers/spree/admin/base_controller.rb
@@ -3,6 +3,7 @@ module Spree
     class BaseController < Spree::BaseController
       helper 'spree/admin/navigation'
       helper 'spree/locale'
+      helper 'spree/currency'
       layout 'spree/layouts/admin'
 
       before_action :authorize_admin
